+++ miamoremustachio/.gitignore
@@ -1,3 +1,2 @@
 **/package-lock.json
-**/package.json
 **/node_modules
