+++ miamoremustachio/app/modules/constants.js
@@ -1,0 +1,10 @@
+module.exports.OPERATIONS = {
+  ADD: 'add',
+  SUBTRACT: 'subtract',
+  MULTI: 'multi',
+};
+
+module.exports.ERRORS = {
+  INVALID_OPERATION: 'Error: Invalid operation.',
+  INVALID_OPERAND: 'Error: Invalid operand.',
+};
