+++ miamoremustachio/app/modules/helpers.js
@@ -1,0 +1,14 @@
+const { ERRORS } = require('./constants');
+
+function isOperandValid(operand) {
+  return Number.isFinite(operand);
+}
+
+module.exports.validateOperand = function(a, b) {
+  if (isOperandValid(a) && isOperandValid(b)) {
+    return true;
+  
+  } else {
+    return console.error(ERRORS.INVALID_OPERAND);
+  }
+}
