+++ miamoremustachio/app/modules/operations.js
@@ -1,0 +1,3 @@
+module.exports.add = (a, b) => a + b;
+module.exports.subtract = (a, b) => a - b;
+module.exports.multiply = (a, b) => a * b;
