+++ miamoremustachio/jsconfig.json
@@ -1,0 +1,7 @@
+{
+  "compilerOptions": {
+    "module": "CommonJS",
+    // "target": "ES6"
+  },
+  "exclude": ["node_modules", "**/node_modules/*"]
+}
