+++ miamoremustachio/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "calc",
+  "version": "1.0.0",
+  "main": "main.js",
+  "devDependencies": {
+    "@types/node": "^22.7.4"
+  },
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "miamoremustachio",
+  "license": "ISC",
+  "description": ""
+}
